clinic_chatbot/requirements.txt
google-generativeai==0.3.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pydantic==2.5.0
python-dateutil==2.8.2